progress

+++ src/customer_backend/src/lib.rs
@@ -5,10 +5,10 @@ use candid::Principal;
 
 /// Calls an arbitrary method on an arbitrary canister with an arbitrary amount of cycles attached.
 #[update()]
-async fn call_with_attached_cycles(call_params: (Principal, String, u128)) -> Result<String, PaymentError> {
+async fn call_with_attached_cycles(call_params: (Principal, String, u64)) -> Result<String, PaymentError> {
     let (canister_id, method, cycles) = call_params;
     let arg = ();
-    let (ans, ): (Result<String, PaymentError>,) = call_with_payment128(canister_id, &method, arg, cycles).await.unwrap();
+    let (ans, ): (Result<String, PaymentError>,) = call_with_payment128(canister_id, &method, arg, u128::from(cycles)).await.unwrap();
     ans
 }
 /*
